Clean up

+++ src/ulb/infof307/g01/server/database/DatabaseSchema.java
@@ -102,7 +102,7 @@ class DatabaseSchema {
             """
         CREATE TABLE IF NOT EXISTS user_deck_score (
             user_id TEXT,
-            timestamp INTEGER(8) NOT NULL,
+            timestamp INTEGER NOT NULL,
             deck_id TEXT NOT NULL,
             score INTEGER NOT NULL,
             PRIMARY KEY (user_id, timestamp),
